Update pruning table for stage 3 edges

+++ PruningEdgCenStage5.java
@@ -12,8 +12,8 @@ public final class PruningEdgCenStage5 extends Pruning {
 		num_moves = Constants.N_SQMOVES;
 
 		// Creation of the pruning table.
-		num_positions = Constants.N_SQS_SYMEDGE_PERM*Constants.N_SQS_CENTER_PERM;
-		int n = num_positions/4 + 1;
+		num_positions = (long)(Constants.N_SQS_SYMEDGE_PERM*Constants.N_SQS_CENTER_PERM);
+		int n = (int)(num_positions/4 + 1);
 		CubeSqsCoord.prune_table_edgcen = new byte[n];
 		ptable = CubeSqsCoord.prune_table_edgcen;
 		for (i = 0; i < n; ++i) {
@@ -25,9 +25,9 @@ public final class PruningEdgCenStage5 extends Pruning {
 		set_dist(21616*Constants.N_SQS_CENTER_PERM+143, 3);
 	}
 
-	int do_move (int idx, int move){
+	long do_move (long idx, int move){
 		short cen = (short)(idx % Constants.N_SQS_CENTER_PERM);
-		int edge = idx / Constants.N_SQS_CENTER_PERM;
+		int edge = (int)(idx / Constants.N_SQS_CENTER_PERM);
 
 		int newEdge = Tables.move_table_symEdgeSTAGE5[edge][move];
 		int sym = newEdge % 48;
@@ -39,10 +39,10 @@ public final class PruningEdgCenStage5 extends Pruning {
 		return edgeRep*Constants.N_SQS_CENTER_PERM + cen;
 	}
 
-	void saveIdxAndSyms (int idx, int dist){
+	void saveIdxAndSyms (long idx, int dist){
 		set_dist (idx, dist);
 
-		int edge = idx / Constants.N_SQS_CENTER_PERM;
+		int edge = (int)(idx / Constants.N_SQS_CENTER_PERM);
 		short cen = (short)(idx % Constants.N_SQS_CENTER_PERM);
 		int symI = 0;
 		long syms = Tables.hasSymEdgeSTAGE5[edge];
